test: BalanceService 동시성 테스트 분리
